Refactored sim and panel modules, added test for new panel
