+++ microhapulator/__init__.py
@@ -10,7 +10,7 @@
 # Core libraries
 import builtins
 from gzip import open as gzopen
-from sys import stdin, stdout, stderr
+import sys
 
 # Internal modules
 from microhapulator import panel
@@ -38,7 +38,7 @@ def open(filename, mode):
     if mode not in ('r', 'w'):
         raise ValueError('invalid mode "{}"'.format(mode))
     if filename in ['-', None]:
-        filehandle = stdin if mode == 'r' else stdout
+        filehandle = sys.stdin if mode == 'r' else sys.stdout
         return filehandle
     openfunc = builtins.open
     if filename.endswith('.gz'):
@@ -52,4 +52,4 @@ def plog(*args, **kwargs):
     if logstream is not None:
         print(*args, **kwargs, file=logstream)
     if logstream is None or teelog:
-        print(*args, **kwargs, file=stderr)
+        print(*args, **kwargs, file=sys.stderr)
